requirement.txt
elevenlabs
json
os
python-dotenv
SpeechRecognition 
pydub 
pyaudio
pyttsx3